avoid converting pdf to pdf; avoid maven test failure due to libreoffice

+++ src/main/java/cn/edu/nju/teamwiki/api/controller/DocumentController.java
@@ -3,13 +3,19 @@ package cn.edu.nju.teamwiki.api.controller;
 import cn.edu.nju.teamwiki.api.Result;
 import cn.edu.nju.teamwiki.api.param.RenameDocumentParams;
 import cn.edu.nju.teamwiki.api.vo.DocumentVO;
+import cn.edu.nju.teamwiki.jooq.tables.pojos.Document;
 import cn.edu.nju.teamwiki.service.DocumentService;
 import cn.edu.nju.teamwiki.service.ServiceException;
+import cn.edu.nju.teamwiki.util.OfficeUtils;
 import cn.edu.nju.teamwiki.util.SessionUtils;
+import cn.edu.nju.teamwiki.util.StorageUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import org.jodconverter.DocumentConverter;
+import org.jodconverter.office.LocalOfficeManager;
+import org.jodconverter.office.OfficeException;
+import org.jodconverter.office.OfficeManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,9 +45,6 @@ public class DocumentController {
 
     @Autowired
     private DocumentService documentService;
-
-    @Autowired
-    private DocumentConverter documentConverter;
 
 
     @GetMapping
@@ -108,13 +111,17 @@ public class DocumentController {
     @ApiOperation("以pdf形式预览文档")
     public ResponseEntity<Resource> preview(@PathVariable("id") String documentId) throws Exception {
         File documentFile = documentService.getDocumentAbsolutePath(documentId).toFile();
-        File previewFile = File.createTempFile(documentId, ".pdf");
-        documentConverter.convert(documentFile).to(previewFile).execute();
-        previewFile.deleteOnExit();
-        Resource resource = new UrlResource(previewFile.toURI());
+        File previewFile;
+        if (!StorageUtils.getFileSuffixName(documentFile).equals("pdf")) {
+            previewFile = documentFile;
+        } else {
+            previewFile = File.createTempFile(StorageUtils.getFilePrefixName(documentFile), ".pdf");
+            OfficeUtils.convert(documentFile, previewFile);
+            previewFile.deleteOnExit();
+        }
         return ResponseEntity.ok()
                 .contentType(MediaType.APPLICATION_PDF)
                 .header(HttpHeaders.CONTENT_DISPOSITION, "inline; filename=" + previewFile.getName())
-                .body(resource);
+                .body(new UrlResource(previewFile.toURI()));
     }
 }
